pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dataviewer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.7",
    "duckdb>=1.4.4",
    "fastapi[standard]>=0.128.8",
    "ibis-framework[duckdb]>=12.0.0",
    "lancedb>=0.29.2",
    "polars>=1.38.1",
    "pydanclick[griffe]>=0.5.1",
    "pydantic-settings>=2.12.0",
    "pylance>=2.0.0",
    "rich>=14.3.2",
    "uvicorn>=0.32.0",
]

[project.scripts]
dataviewer = "dataviewer.cli:main"

[tool.setuptools.packages.find]
include = ["dataviewer*"]

[tool.pyright]
include = ["dataviewer"]
pythonVersion = "3.12"
typeCheckingMode = "basic"
reportMissingTypeStubs = false

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "ANN"]
ignore = ["E203"]

[tool.ruff.lint.isort]
known-first-party = ["dataviewer"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[dependency-groups]
dev = [
    "huggingface-hub[hf-transfer]>=1.4.1",
    "ipywidgets>=8.1.8",
    "jupyterlab>=4.5.4",
    "pre-commit>=3.7.0",
    "pyright==1.1.408",
    "pytest>=9.0.2",
    "ruff>=0.6.0",
]
